separate InstanceSerializer and InstanceDeserializer from InstanceParser
